+++ src/_pytest/deprecated.py
@@ -48,16 +48,6 @@ RESULT_LOG = PytestDeprecationWarning(
 )
 
 
-PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST = (
-    "Defining 'pytest_plugins' in a non-top-level conftest is no longer supported "
-    "because it affects the entire directory tree in a non-explicit way.\n"
-    "  {}\n"
-    "Please move it to a top level conftest file at the rootdir:\n"
-    "  {}\n"
-    "For more information, visit:\n"
-    "  https://docs.pytest.org/en/latest/deprecations.html#pytest-plugins-in-non-top-level-conftest-files"
-)
-
 PYTEST_CONFIG_GLOBAL = PytestDeprecationWarning(
     "the `pytest.config` global is deprecated.  Please use `request.config` "
     "or `pytest_configure` (if you're a pytest plugin) instead."
