+++ testing/deprecated_test.py
@@ -1,8 +1,5 @@
-import os
-
 import pytest
 from _pytest import deprecated
-from _pytest.warnings import SHOW_PYTEST_WARNINGS_ARG
 
 pytestmark = pytest.mark.pytester_example_path("deprecated")
 
@@ -73,121 +70,6 @@ def test_external_plugins_integrated(testdir, plugin):
 
     with pytest.warns(pytest.PytestConfigWarning):
         testdir.parseconfig("-p", plugin)
-
-
-def test_pytest_plugins_in_non_top_level_conftest_deprecated(testdir):
-    from _pytest.deprecated import PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST
-
-    testdir.makepyfile(
-        **{
-            "subdirectory/conftest.py": """
-        pytest_plugins=['capture']
-    """
-        }
-    )
-    testdir.makepyfile(
-        """
-        def test_func():
-            pass
-    """
-    )
-    res = testdir.runpytest(SHOW_PYTEST_WARNINGS_ARG)
-    assert res.ret == 2
-    msg = str(PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST).splitlines()[0]
-    res.stdout.fnmatch_lines(
-        ["*{msg}*".format(msg=msg), "*subdirectory{sep}conftest.py*".format(sep=os.sep)]
-    )
-
-
-@pytest.mark.parametrize("use_pyargs", [True, False])
-def test_pytest_plugins_in_non_top_level_conftest_unsupported_pyargs(
-    testdir, use_pyargs
-):
-    """When using --pyargs, do not emit the warning about non-top-level conftest warnings (#4039, #4044)"""
-    from _pytest.deprecated import PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST
-
-    files = {
-        "src/pkg/__init__.py": "",
-        "src/pkg/conftest.py": "",
-        "src/pkg/test_root.py": "def test(): pass",
-        "src/pkg/sub/__init__.py": "",
-        "src/pkg/sub/conftest.py": "pytest_plugins=['capture']",
-        "src/pkg/sub/test_bar.py": "def test(): pass",
-    }
-    testdir.makepyfile(**files)
-    testdir.syspathinsert(testdir.tmpdir.join("src"))
-
-    args = ("--pyargs", "pkg") if use_pyargs else ()
-    args += (SHOW_PYTEST_WARNINGS_ARG,)
-    res = testdir.runpytest(*args)
-    assert res.ret == (0 if use_pyargs else 2)
-    msg = str(PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST).splitlines()[0]
-    if use_pyargs:
-        assert msg not in res.stdout.str()
-    else:
-        res.stdout.fnmatch_lines(["*{msg}*".format(msg=msg)])
-
-
-def test_pytest_plugins_in_non_top_level_conftest_unsupported_no_top_level_conftest(
-    testdir
-):
-    from _pytest.deprecated import PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST
-
-    subdirectory = testdir.tmpdir.join("subdirectory")
-    subdirectory.mkdir()
-    testdir.makeconftest(
-        """
-        pytest_plugins=['capture']
-    """
-    )
-    testdir.tmpdir.join("conftest.py").move(subdirectory.join("conftest.py"))
-
-    testdir.makepyfile(
-        """
-        def test_func():
-            pass
-    """
-    )
-
-    res = testdir.runpytest_subprocess()
-    assert res.ret == 2
-    msg = str(PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST).splitlines()[0]
-    res.stdout.fnmatch_lines(
-        ["*{msg}*".format(msg=msg), "*subdirectory{sep}conftest.py*".format(sep=os.sep)]
-    )
-
-
-def test_pytest_plugins_in_non_top_level_conftest_unsupported_no_false_positives(
-    testdir
-):
-    from _pytest.deprecated import PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST
-
-    subdirectory = testdir.tmpdir.join("subdirectory")
-    subdirectory.mkdir()
-    testdir.makeconftest(
-        """
-        pass
-    """
-    )
-    testdir.tmpdir.join("conftest.py").move(subdirectory.join("conftest.py"))
-
-    testdir.makeconftest(
-        """
-        import warnings
-        warnings.filterwarnings('always', category=DeprecationWarning)
-        pytest_plugins=['capture']
-    """
-    )
-    testdir.makepyfile(
-        """
-        def test_func():
-            pass
-    """
-    )
-    res = testdir.runpytest_subprocess()
-    assert res.ret == 0
-    msg = str(PYTEST_PLUGINS_FROM_NON_TOP_LEVEL_CONFTEST).splitlines()[0]
-    assert msg not in res.stdout.str()
 
 
 def test_fixture_named_request(testdir):
