+++ testing/test_config.py
@@ -1,3 +1,4 @@
+import os
 import sys
 import textwrap
 
@@ -1242,3 +1243,116 @@ def test_config_blocked_default_plugins(testdir, plugin):
         result.stdout.fnmatch_lines(["* 1 failed in *"])
     else:
         assert result.stdout.lines == [""]
+
+
+class TestPytestPluginsVariable:
+    def test_pytest_plugins_in_non_top_level_conftest_unsupported(self, testdir):
+        testdir.makepyfile(
+            **{
+                "subdirectory/conftest.py": """
+            pytest_plugins=['capture']
+        """
+            }
+        )
+        testdir.makepyfile(
+            """
+            def test_func():
+                pass
+        """
+        )
+        res = testdir.runpytest()
+        assert res.ret == 2
+        msg = "Defining 'pytest_plugins' in a non-top-level conftest is no longer supported"
+        res.stdout.fnmatch_lines(
+            [
+                "*{msg}*".format(msg=msg),
+                "*subdirectory{sep}conftest.py*".format(sep=os.sep),
+            ]
+        )
+
+    @pytest.mark.parametrize("use_pyargs", [True, False])
+    def test_pytest_plugins_in_non_top_level_conftest_unsupported_pyargs(
+        self, testdir, use_pyargs
+    ):
+        """When using --pyargs, do not emit the warning about non-top-level conftest warnings (#4039, #4044)"""
+
+        files = {
+            "src/pkg/__init__.py": "",
+            "src/pkg/conftest.py": "",
+            "src/pkg/test_root.py": "def test(): pass",
+            "src/pkg/sub/__init__.py": "",
+            "src/pkg/sub/conftest.py": "pytest_plugins=['capture']",
+            "src/pkg/sub/test_bar.py": "def test(): pass",
+        }
+        testdir.makepyfile(**files)
+        testdir.syspathinsert(testdir.tmpdir.join("src"))
+
+        args = ("--pyargs", "pkg") if use_pyargs else ()
+        res = testdir.runpytest(*args)
+        assert res.ret == (0 if use_pyargs else 2)
+        msg = (
+            msg
+        ) = "Defining 'pytest_plugins' in a non-top-level conftest is no longer supported"
+        if use_pyargs:
+            assert msg not in res.stdout.str()
+        else:
+            res.stdout.fnmatch_lines(["*{msg}*".format(msg=msg)])
+
+    def test_pytest_plugins_in_non_top_level_conftest_unsupported_no_top_level_conftest(
+        self, testdir
+    ):
+        subdirectory = testdir.tmpdir.join("subdirectory")
+        subdirectory.mkdir()
+        testdir.makeconftest(
+            """
+            pytest_plugins=['capture']
+        """
+        )
+        testdir.tmpdir.join("conftest.py").move(subdirectory.join("conftest.py"))
+
+        testdir.makepyfile(
+            """
+            def test_func():
+                pass
+        """
+        )
+
+        res = testdir.runpytest_subprocess()
+        assert res.ret == 2
+        msg = "Defining 'pytest_plugins' in a non-top-level conftest is no longer supported"
+        res.stdout.fnmatch_lines(
+            [
+                "*{msg}*".format(msg=msg),
+                "*subdirectory{sep}conftest.py*".format(sep=os.sep),
+            ]
+        )
+
+    def test_pytest_plugins_in_non_top_level_conftest_unsupported_no_false_positives(
+        self, testdir
+    ):
+        subdirectory = testdir.tmpdir.join("subdirectory")
+        subdirectory.mkdir()
+        testdir.makeconftest(
+            """
+            pass
+        """
+        )
+        testdir.tmpdir.join("conftest.py").move(subdirectory.join("conftest.py"))
+
+        testdir.makeconftest(
+            """
+            import warnings
+            warnings.filterwarnings('always', category=DeprecationWarning)
+            pytest_plugins=['capture']
+        """
+        )
+        testdir.makepyfile(
+            """
+            def test_func():
+                pass
+        """
+        )
+        res = testdir.runpytest_subprocess()
+        assert res.ret == 0
+        msg = "Defining 'pytest_plugins' in a non-top-level conftest is no longer supported"
+        assert msg not in res.stdout.str()
